# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Auction API
  version: 1.0.0
  description: |
    Hono REST API for sales, lots, bids, payments, submissions, users, and admin operations.
    Authentication is handled by Better Auth at `/api/auth/*`.
servers:
  - url: http://localhost:8787
paths:
...
  /payments/{id}/capture:
    parameters:
      - $ref: "#/components/parameters/PaymentId"
    post:
      summary: Mark payment captured manually (admin)
  /payments/{id}/refund:
    parameters:
      - $ref: "#/components/parameters/PaymentId"
    post:
      summary: Mark payment refunded manually (admin; DB status only until a gateway is integrated)
...
components:
  parameters:
...
    PaymentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...